fixed help menu

+++ src/org/academiadecodigo/murlogs/castleinvaders/Menu/Menu.java
@@ -159,13 +159,15 @@ public class Menu implements KeyboardHandler {
                 return;
 
             case KeyboardEvent.KEY_ENTER:
-                if (selecter.getY() == 360) {
-                    deleteMenu();
+                if (menuExists && !helpSection.helpIsCreated()) {
+                    if (selecter.getY() == 360) {
+                        deleteMenu();
+                        return;
+                    }
+                    helpSection.createHelpSection();
                     return;
                 }
-                helpSection.createHelpSection();
                 return;
-
             case KeyboardEvent.KEY_ESC:
                 if (!helpSection.helpIsCreated()) {
                     return;
